Remove dom-accessibility-api and use WhatSock dependency instead
The algorithm in `dom-accessibility-api` is incorrectly interpreting `title` as an accessible name.
This PR swaps it out with the algorithm provided in https://github.com/whatsock/w3c-alternative-text-computation
which seems to reflect more of what we want.

The caveat is that there is no NPM package. Therefore I had to create a vendors file and import it manually.